[CIDMDEV-5134] - Migration Support - Read-Only Policies
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :

1. (ADD) POST v2.0/RAX-AUTH/policies
2. (UPDATE) PUT v2.0/RAX-AUTH/policies/{policyId}
3. (DELETE) DELETE v2.0/RAX-AUTH/policies/{policyId}
4. (Add policy to endpoint template) PUT v2.0/OS-KSCATALOG/endpointTemplates/{endpointTemplateId}/RAX-AUTH/policies/{policyId}
5. (Add policy to endpoint template) DELETE v2.0/OS-KSCATALOG/endpointTemplates/{endpointTemplateId}/RAX-AUTH/policies/{policyId}
6. (update policies on endpoint template) PUT v2.0/OS-KSCATALOG/endpointTemplates/{endpointTemplateId}/RAX-AUTH/policies

This prevents modifying endpoint templates w/ respect to associating policy IDs to them. While strictly speaking this is a modification of an Endpoint (CloudBaseURL), since it's dealing with associating a policy to the endpoint, the change is being made as part of the PR.

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -264,7 +264,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @Context UriInfo uriInfo,
             Policy policy) {
-        return cloud20Service.addPolicy(uriInfo, authToken, policy).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addPolicy(uriInfo, authToken, policy).build();
+        }
     }
 
     @GET
@@ -281,7 +285,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("policyId") String policyId,
             Policy policy) {
-        return cloud20Service.updatePolicy(authToken, policyId, policy).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updatePolicy(authToken, policyId, policy).build();
+        }
     }
 
     @DELETE
@@ -289,7 +297,11 @@ public class Cloud20VersionResource {
     public Response deletePolicy(
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("policyId") String policyId) {
-        return cloud20Service.deletePolicy(authToken, policyId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deletePolicy(authToken, policyId).build();
+        }
     }
 
     @GET
@@ -831,7 +843,11 @@ public class Cloud20VersionResource {
             @PathParam("endpointTemplateId") String endpointTemplateId,
             Policies policies
             ) {
-        return cloud20Service.updatePoliciesForEndpointTemplate(authToken, endpointTemplateId, policies).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updatePoliciesForEndpointTemplate(authToken, endpointTemplateId, policies).build();
+        }
     }
 
     @PUT
@@ -840,7 +856,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("endpointTemplateId") String endpointTemplateId,
             @PathParam("policyId") String policyId){
-        return cloud20Service.addPolicyToEndpointTemplate(authToken, endpointTemplateId, policyId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addPolicyToEndpointTemplate(authToken, endpointTemplateId, policyId).build();
+        }
     }
 
     @DELETE
@@ -849,7 +869,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("endpointTemplateId") String endpointTemplateId,
             @PathParam("policyId") String policyId){
-        return cloud20Service.deletePolicyToEndpointTemplate(authToken, endpointTemplateId, policyId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deletePolicyToEndpointTemplate(authToken, endpointTemplateId, policyId).build();
+        }
     }
 
     @GET
